filip(fix): force re-render to initialise scrollY framer motion hooks
